# fynz-srv/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Fynz API
  version: 1.0.0
  description: |
    ## 💰 Fynz — Calculadora Inteligente de Gastos

    API REST para la aplicación Fynz, un sistema de gestión financiera personal
    que permite a los usuarios:

    - **Registrar transacciones** (ingresos, gastos y transferencias)
    - **Organizar dinero en bolsillos** de ahorro con metas
    - **Clasificar movimientos** con categorías personalizables
    - **Consultar resúmenes financieros** con desglose por categoría

    ### Conceptos Clave

    | Concepto | Descripción |
    |----------|-------------|
    | **Balance General** | Calculado dinámicamente: `SUM(income) - SUM(expense) - SUM(transfer)` |
    | **Bolsillos (Pockets)** | Cuentas virtuales de ahorro con metas. Se alimentan con transacciones tipo `transfer` |
    | **Categorías** | Clasificadores para transacciones. Hay 10 del sistema + personalizadas por usuario |
    | **Tipos de Transacción** | `income` (ingreso), `expense` (gasto), `transfer` (transferencia a bolsillo) |

    ### Autenticación

    Todos los endpoints (excepto `/auth/register` y `/auth/login`) requieren un
    **JWT token** en el header `Authorization`:

    ```
    Authorization: Bearer <token>
    ```

    El token se obtiene al hacer login y expira en 1 día (configurable).
  contact:
    name: Isaac — Fynz Dev
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo local

...
# TAGS
...
paths:
...
  /categories:
    get:
      tags: [Categorías]
      summary: Listar todas las categorías
      description: |
        Devuelve las categorías visibles para el usuario autenticado:
        - **Categorías del sistema** (is_default = 1): Alimentación, Transporte, etc.
        - **Categorías personalizadas** del usuario (is_default = 0)

        Ordenadas alfabéticamente por nombre.
      operationId: listCategories
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lista de categorías
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
              example:
                success: true
                data:
                  - id: 1
                    name: "Alimentación"
                    icon: "🍔"
                    color: "#ef4444"
                    user_id: null
                    is_default: 1
                  - id: 11
                    name: "Suscripciones"
                    icon: "📺"
                    color: "#7c3aed"
                    user_id: 1
                    is_default: 0
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    post:
      tags: [Categorías]
      summary: Crear categoría personalizada
      description: |
        Crea una nueva categoría personalizada para el usuario.
        No se permite duplicar nombres de categorías del sistema ni propias.
      operationId: createCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
            examples:
              con_icono:
                summary: Con ícono y color
                value:
                  name: "Suscripciones"
                  icon: "📺"
                  color: "#7c3aed"
              solo_nombre:
                summary: Solo nombre (color default)
                value:
                  name: "Mascotas"
      responses:
        "201":
          description: Categoría creada
          content:
            application/json:
              example:
                success: true
                message: "Categoría creada"
                data:
                  id: 11
                  name: "Suscripciones"
                  icon: "📺"
                  color: "#7c3aed"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          description: Ya existe una categoría con ese nombre
          content:
            application/json:
              example:
                success: false
                message: "Ya existe una categoría con ese nombre"
  /categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID de la categoría
        schema:
          type: integer
          example: 11

    put:
      tags: [Categorías]
      summary: Actualizar categoría personalizada
      description: |
        Actualiza una categoría personalizada del usuario.
        **No se pueden editar categorías del sistema** (is_default = 1).
        Se requiere al menos un campo para actualizar.
      operationId: updateCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
            example:
              name: "Streamings"
              icon: "🎬"
      responses:
        "200":
          description: Categoría actualizada
          content:
            application/json:
              example:
                success: true
                message: "Categoría actualizada"
                data:
                  id: 11
                  name: "Streamings"
                  icon: "🎬"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: No tienes permiso (categoría del sistema o de otro usuario)
          content:
            application/json:
              example:
                success: false
                message: "No puedes editar categorías del sistema"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      tags: [Categorías]
      summary: Eliminar categoría personalizada
      description: |
        Elimina una categoría personalizada del usuario.
        **No se pueden eliminar categorías del sistema.**
        Las transacciones asociadas conservarán category_id = NULL.
      operationId: deleteCategory
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Categoría eliminada
          content:
            application/json:
              example:
                success: true
                message: "Categoría eliminada"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: No tienes permiso para eliminar esta categoría
          content:
            application/json:
              example:
                success: false
                message: "No puedes eliminar categorías del sistema"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ─────────────────────────────────────────────────────────
  # POCKETS
  # ─────────────────────────────────────────────────────────
  /pockets:
    get:
      tags: [Bolsillos]
      summary: Listar bolsillos del usuario
      description: |
        Devuelve todos los bolsillos del usuario autenticado con:
        - **current_amount**: saldo calculado dinámicamente (SUM de transfers)
        - **progress**: porcentaje de avance hacia la meta

        Ordenados por fecha de creación (más recientes primero).
      operationId: listPockets
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lista de bolsillos con saldos calculados
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Pocket"
              example:
                success: true
                data:
                  - id: 1
                    name: "Vacaciones Cancún"
                    target_amount: 25000.00
                    current_amount: 8500.00
                    color: "#3b82f6"
                    icon: "🏖️"
                    progress: 34.0
                    user_id: 1
                    created_at: "2026-03-10T23:00:00.000Z"
                  - id: 2
                    name: "Fondo de Emergencia"
                    target_amount: 50000.00
                    current_amount: 15000.00
                    color: "#10b981"
                    icon: "🛡️"
                    progress: 30.0
                    user_id: 1
                    created_at: "2026-03-08T10:00:00.000Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    post:
      tags: [Bolsillos]
      summary: Crear nuevo bolsillo
      description: |
        Crea un bolsillo de ahorro para organizar dinero por metas.
        El saldo inicial siempre es 0 — se incrementa con transacciones
        de tipo `transfer`.
      operationId: createPocket
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePocketRequest"
            examples:
              con_meta:
                summary: Con meta de ahorro
                value:
                  name: "Auto nuevo"
                  target_amount: 180000.00
                  color: "#f59e0b"
                  icon: "🚗"
              basico:
                summary: Sin meta (solo organizar)
                value:
                  name: "Gastos compartidos"
      responses:
        "201":
          description: Bolsillo creado
          content:
            application/json:
              example:
                success: true
                message: "Bolsillo creado"
                data:
                  id: 1
                  name: "Auto nuevo"
                  target_amount: 180000.00
                  color: "#f59e0b"
                  icon: "🚗"
                  current_amount: 0
                  progress: 0
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /pockets/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID del bolsillo
        schema:
          type: integer
          example: 1

    get:
      tags: [Bolsillos]
      summary: Obtener un bolsillo por ID
      description: |
        Devuelve los detalles de un bolsillo específico del usuario,
        incluyendo su saldo actual y progreso calculados dinámicamente.
      operationId: getPocketById
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Detalle del bolsillo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Pocket"
              example:
                success: true
                data:
                  id: 1
                  name: "Vacaciones Cancún"
                  target_amount: 25000.00
                  current_amount: 8500.00
                  color: "#3b82f6"
                  icon: "🏖️"
                  progress: 34.0
                  user_id: 1
                  created_at: "2026-03-10T23:00:00.000Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      tags: [Bolsillos]
      summary: Actualizar bolsillo
      description: |
        Actualiza los datos de un bolsillo existente.
        Se requiere al menos un campo para actualizar.
        **No modifica las transacciones asociadas.**
      operationId: updatePocket
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePocketRequest"
            example:
              name: "Vacaciones Europa 2027"
              target_amount: 60000.00
      responses:
        "200":
          description: Bolsillo actualizado
          content:
            application/json:
              example:
                success: true
                message: "Bolsillo actualizado"
                data:
                  id: 1
                  name: "Vacaciones Europa 2027"
                  target_amount: 60000.00
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      tags: [Bolsillos]
      summary: Eliminar bolsillo
      description: |
        Elimina un bolsillo del usuario. Las transacciones asociadas
        conservarán su registro pero con pocket_id = NULL (ON DELETE SET NULL).
      operationId: deletePocket
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Bolsillo eliminado
          content:
            application/json:
              example:
                success: true
                message: "Bolsillo eliminado"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ─────────────────────────────────────────────────────────
  # TRANSACTIONS
  # ─────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido del endpoint `/auth/login`.
        Header: `Authorization: Bearer <token>`

  # ─── Schemas ────────────────────────────────────────────
  schemas:
...
    Category:
      type: object
      description: |
        Categoría para clasificar transacciones.
        is_default = 1: categoría del sistema (no editable).
        is_default = 0: categoría personalizada del usuario.
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 50
          example: "Alimentación"
        icon:
          type: string
          nullable: true
          example: "🍔"
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          example: "#ef4444"
        user_id:
          type: integer
          nullable: true
          description: "NULL = categoría global del sistema"
          example: null
        is_default:
          type: integer
          enum: [0, 1]
          description: "1 = sistema (no editable), 0 = personalizada"
          example: 1
      required: [id, name]
    CreateCategoryRequest:
      type: object
      description: Datos para crear una categoría personalizada
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: "Suscripciones"
        icon:
          type: string
          maxLength: 10
          example: "📺"
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          example: "#7c3aed"
      required: [name]
    UpdateCategoryRequest:
      type: object
      description: Datos para actualizar categoría (mínimo 1 campo)
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: "Streamings"
        icon:
          type: string
          maxLength: 10
          example: "🎬"
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          example: "#a855f7"

    # ─── Pocket ───────────────────────────────────────────
    Pocket:
      type: object
      description: |
        Bolsillo de ahorro con meta.
        current_amount y progress se calculan dinámicamente (no se almacenan).
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 50
          example: "Vacaciones Cancún"
        target_amount:
          type: number
          format: double
          minimum: 0
          description: Meta de ahorro en pesos
          example: 25000.00
        current_amount:
          type: number
          format: double
          readOnly: true
          description: "Saldo = SUM(transactions WHERE pocket_id = this.id AND type = 'transfer')"
          example: 8500.00
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          example: "#3b82f6"
        icon:
          type: string
          nullable: true
          example: "🏖️"
        progress:
          type: number
          format: double
          readOnly: true
          minimum: 0
          maximum: 100
          description: "(current_amount / target_amount) * 100"
          example: 34.0
        user_id:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2026-03-10T23:00:00.000Z"
      required: [id, name, user_id]
    CreatePocketRequest:
      type: object
      description: Datos para crear un nuevo bolsillo
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: "Fondo de Emergencia"
        target_amount:
          type: number
          format: double
          minimum: 0
          default: 0
          example: 50000.00
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          example: "#10b981"
        icon:
          type: string
          maxLength: 10
          example: "🛡️"
      required: [name]
    UpdatePocketRequest:
      type: object
      description: Datos para actualizar bolsillo (mínimo 1 campo)
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: "Fondo de Emergencia 2026"
        target_amount:
          type: number
          format: double
          minimum: 0
          example: 75000.00
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          example: "#059669"
        icon:
          type: string
          maxLength: 10
          example: "💰"

    # ─── Transaction ──────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Recurso no encontrado"
    ErrorValidation:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Datos inválidos"
        errors:
          type: object
          description: Errores por campo (formato Zod)
          additionalProperties:
            type: object
            properties:
              _errors:
                type: array
                items:
                  type: string

  # ─── Respuestas reutilizables ───────────────────────────
  responses:
    UnauthorizedError:
      description: Token JWT faltante, inválido o expirado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: "Token inválido o expirado"
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: "Recurso no encontrado"
    ValidationError:
      description: Error de validación de datos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorValidation"
...
security:
  - BearerAuth: []